flask_sentiment_api/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch 
torchvision 
torchaudio 
Flask==2.2.2
Flask_Cors==3.0.10
grpcio-tools
git+https://github.com/huggingface/transformers.git